+++ tests/test_function_type.py
@@ -1,10 +1,9 @@
 from typing import TYPE_CHECKING
 
-from basedtyping.callables import Function
-from basedtyping.type_checks import assert_type
-
 if TYPE_CHECKING:
-    # these are just type-time tests, not real life pytest tests. they are only run by mypy
+
+    from basedtyping import Function
+    from basedtyping.typetime_only import assert_type
 
     assert_function = assert_type[Function]
