+++ tests/test_typetime_only/test_assert_type.py
@@ -1,9 +1,9 @@
 from typing import TYPE_CHECKING
 
-from basedtyping.type_checks import assert_type
-
 if TYPE_CHECKING:
     # these are just type-time tests, not real life pytest tests. they are only run by mypy
+    from basedtyping.typetime_only import assert_type
+
     def test_assert_type_pass() -> None:
         assert_type[int](1)
