+++ tests/test_typetime_only/test_typetime_only.py
@@ -1,0 +1,6 @@
+from pytest import raises
+
+
+def test_runtime_import() -> None:
+    with raises(ImportError):
+        import basedtyping.typetime_only  # pylint:disable=import-outside-toplevel # noqa: F401
